Remove "task_uuid" attribute in HeavyCompuation model
"task_uuid" attribute was put in the HeavyComputation model in order to
trak the progress of the linked Celery task.
However, the code that write the "task_uuid" is in the @shared_task
definition. This implies that the code is only executed after the task
has reach the "STARTED" Celery state. Before that moment (ex. while the
task is in the "RECEIVED" state) the "task_uuid" field remains empty.
This behaviour creates confusion as one cannot rely on the "task_uuid"
to track the Celery task from start to end.